# Conflicts:
#	data/assets/examples/luastatemachine.asset
#	data/assets/examples/statemachine.asset
#	modules/statemachine/include/state.h
#	modules/statemachine/include/statemachine.h
#	modules/statemachine/include/transition.h
#	modules/statemachine/src/state.cpp
#	modules/statemachine/src/statemachine.cpp
#	modules/statemachine/src/transition.cpp
#	modules/statemachine/statemachinemodule.cpp
#	modules/statemachine/statemachinemodule.h

+++ modules/debugging/debuggingmodule_lua.inl
@@ -245,7 +245,7 @@ int renderPathControlPoints(lua_State* L) {
         "openspace.createSingleColorImage('point_color', { 0.0, 1.0, 0.0 })"
     ")";
 
-    for (int i = 0; i < points.size(); i++) {
+    for (size_t i = 0; i < points.size(); i++) {
         const std::string& node = "{"
             "Identifier = 'ControlPoint_" + std::to_string(i) + "',"
             "Parent = '" + RenderedPointsIdentifier + "',"
